feat: add new home tab
