+++ src/pages/proposal.tsx
@@ -51,8 +51,10 @@ import useProposalUpdates from '../hooks/useProposalUpdates'
 import locations from '../modules/locations'
 import { isUnderMaintenance } from '../modules/maintenance'
 
-import './index.css'
 import './proposal.css'
+import './proposals.css'
+
+// TODO: Review why proposals.css is being imported and use only proposal.css
 
 type ProposalPageOptions = {
   changing: boolean
