+++ src/pages/proposals.css
